requirements.txt
tensorflow>=2.4.0
keras>=2.4.0
numpy>=1.19.0
pandas>=1.1.0
matplotlib>=3.3.0
scikit-learn>=0.23.0
opencv-python>=4.4.0
tqdm>=4.50.0
pillow>=8.0.0
seaborn>=0.11.0
scipy>=1.5.0 